+++ components/bsx/Create/CreateToken.vue
@@ -81,7 +81,7 @@
 </template>
 
 <script lang="ts">
-import { offsetAttribute } from '@/components/rmrk/Create/mintUtils'
+import { offsetAttribute } from '@/utils/mintUtils'
 import ChainMixin from '@/utils/mixins/chainMixin'
 import { notificationTypes, showNotification } from '@/utils/notification'
 import shouldUpdate from '@/utils/shouldUpdate'
